# Binary quantization using bit shifts.