--- a/adapters/repos/db/vector/compressionhelpers/binary_quantization.go
+++ b/adapters/repos/db/vector/compressionhelpers/binary_quantization.go
@@ -12,8 +12,6 @@
 package compressionhelpers
 
 import (
-	"math"
-
 	"github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer"
 )
 
@@ -36,7 +34,7 @@ func (bq BinaryQuantizer) Encode(vec []float32) []uint64 {
 	for j := 0; j < len(vec); j++ {
 		if vec[j] < 0 {
 			segment := j / 64
-			code[segment] += uint64(math.Pow(2, float64(j%64)))
+			code[segment] |= uint64(1) << (j % 64)
 		}
 	}
 	return code