--- a/adapters/repos/db/vector/compressionhelpers/binary_quantization_test.go
+++ b/adapters/repos/db/vector/compressionhelpers/binary_quantization_test.go
@@ -13,6 +13,8 @@ package compressionhelpers_test
 
 import (
 	"fmt"
+	"math"
+	"math/rand"
 	"sync"
 	"testing"
 	"time"
@@ -87,4 +89,82 @@ func TestBinaryQuantizerChecksSize(t *testing.T) {
 	bq := compressionhelpers.NewBinaryQuantizer(nil)
 	_, err := bq.DistanceBetweenCompressedVectors(make([]uint64, 3), make([]uint64, 4))
 	assert.NotNil(t, err)
+}
+
+func extractBit(code []uint64, idx int) bool {
+	return code[idx/64]&(uint64(1)<<(idx%64)) != 0
+}
+
+func TestBinaryQuantizerBitAssignmenFixedValues(t *testing.T) {
+	test_bits := []struct {
+		value           float32
+		quantized_value bool
+	}{
+		{-1.0, true},
+		{1.0, false},
+		{0.0, false},
+		{float32(math.NaN()), false},
+		{float32(math.Inf(1)), false},
+		{float32(math.Inf(-1)), true},
+	}
+
+	bq := compressionhelpers.NewBinaryQuantizer(nil)
+	for _, b := range test_bits {
+		y := []float32{b.value}
+		code := bq.Encode(y)
+		assert.True(t, extractBit(code, 0) == b.quantized_value,
+			"Unexpected quantized value: %f should quantize to %t",
+			b.value, b.quantized_value)
+	}
+}
+
+func TestBinaryQuantizerBitAssignmenRandomVector(t *testing.T) {
+	const seed = 42
+	r := rand.New(rand.NewSource(seed))
+	x := make([]float32, 1000)
+	for i := range len(x) {
+		x[i] = 2.0*r.Float32() - 1.0
+	}
+
+	bq := compressionhelpers.NewBinaryQuantizer(nil)
+	code := bq.Encode(x)
+	for i, v := range x {
+		assert.True(t, extractBit(code, i) == (v < 0),
+			"Unexpected quantized value: %f should quantize to %t", v, v < 0)
+	}
+
+	for i := 1000; i < 1024; i++ {
+		assert.True(t, extractBit(code, i) == false,
+			"Remaining bits should be set to zero.")
+	}
+}
+
+// Verify that using bit shifts produces the same results as the previous
+// approach of computing powers of floats and converting to uint64.
+func TestBinaryQuantizerBitShiftBackwardsCompatibility(t *testing.T) {
+	for j := range 64 {
+		pow_bit := uint64(math.Pow(2, float64(j%64)))
+		shift_bit := uint64(1) << (j % 64)
+		assert.True(t, pow_bit == shift_bit)
+	}
+}
+
+func BenchmarkBinaryQuantization(b *testing.B) {
+	bq := compressionhelpers.NewBinaryQuantizer(nil)
+	const seed = 42
+	r := rand.New(rand.NewSource(seed))
+	for _, d := range []int{32, 64, 100, 256, 500, 1024, 4096} {
+		x := make([]float32, d)
+		for i := range d {
+			x[i] = 2.0*r.Float32() - 1.0
+		}
+		b.Run(fmt.Sprintf("BinaryQuantization-dim-%d", d), func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				// Even though we do not use the output of bq.Encode() the call
+				// does not seem to be optimized away by the compiler.
+				// TODO: Use b.Loop() instead when we move to Go 1.24.
+				bq.Encode(x)
+			}
+		})
+	}
 }