This patch changes the Encode() function of binary quantization to set the individual bits of the quantized code using bit shifts. Prior to this patch we used the less efficient math.Pow() to produce a desired power of two as a floating point number before converting the result to an uint64 in order to set a specific bit in the quantized code.

A benchmark on random vectors of different lengths shows that using bit shifts instead of math.Pow() results in a 6-8x speedup of Encode().

This patch also includes unit tests to verify that the behavior of binary quantization has not changed with this patch, i.e. we still turn floating point values into bits in exactly the same way.

The output from benchstat for BenchmarkBinaryQuantization comparing the using math.Pow() (old.txt) vs bit shifts (new.txt):

```
goos: darwin
goarch: arm64
cpu: Apple M4 Pro

            │   old.txt    │               new.txt               │
            │    sec/op    │   sec/op     vs base                │
dim-32-14     175.50n ± 2%   20.45n ± 1%  -88.35% (p=0.000 n=10)
dim-64-14     270.60n ± 1%   42.04n ± 1%  -84.46% (p=0.000 n=10)
dim-100-14    534.20n ± 1%   70.65n ± 3%  -86.78% (p=0.000 n=10)
dim-256-14    1444.5n ± 1%   202.1n ± 0%  -86.01% (p=0.000 n=10)
dim-500-14    2746.5n ± 1%   391.0n ± 2%  -85.76% (p=0.000 n=10)
dim-1024-14   5575.0n ± 1%   828.2n ± 1%  -85.14% (p=0.000 n=10)
dim-4096-14   23.590µ ± 1%   3.418µ ± 1%  -85.51% (p=0.000 n=10)
geomean        1.446µ        201.7n       -86.05%
```

### Review checklist
* [x]  Documentation has been updated, if necessary. _Not necessary, pure (minor) performance improvement._
* [x]  Chaos pipeline run or not necessary. _Not necessary, the change is covered by unit tests._
* [x]  All new code is covered by tests where it is reasonable.
* [x]  Performance tests have been run or not necessary. _The specific change has been benchmarked._